+++ src/Lexer/Lexer.php
@@ -75,13 +75,24 @@ final class Lexer
             $type = TokenType::lookupIdentifier($ident);
             $token = new Token($type, $ident);
         } elseif ($this->isNumber($this->char)) {
-            $token = new Token(TokenType::INTEGER, $this->readNumber());
+            $num = $this->readNumber();
+            $token = new Token($this->readNumberTokenType($num), $num);
         } else {
             $token = Token::illegal($this->char);
             $this->advanceChar();
         }
 
         return $token;
+    }
+
+    private function readNumberTokenType(string $num): TokenType
+    {
+        // If number contains more then one dot, the token is ILLEGAL
+        if (substr_count($num, '.') > 1) {
+            return TokenType::ILLEGAL;
+        }
+
+        return str_contains($num, '.') ? TokenType::FLOAT : TokenType::INTEGER;
     }
 
     private function readHtmlToken(): Token
@@ -133,20 +144,34 @@ final class Lexer
         return preg_match('/[_a-zA-Z]/', $letter) === 1;
     }
 
-    private function isNumber(int|string $number): bool
+    private function isInteger(int|string $number): bool
     {
         if ($number === 0) {
+            return false;
+        }
+
+        // The number must not contain a dot
+        if (str_contains((string) $number, '.')) {
             return false;
         }
 
         return preg_match('/[0-9]/', $number) === 1;
     }
 
+    private function isNumber(string $number): bool
+    {
+        if ($number === 0) {
+            return false;
+        }
+
+        return preg_match('/[0-9.]/', $number) === 1;
+    }
+
     private function readIdentifier(): string
     {
         $position = $this->position;
 
-        while ($this->isLetter($this->char) || $this->isNumber($this->char)) {
+        while ($this->isLetter($this->char) || $this->isInteger($this->char)) {
             $this->advanceChar();
         }
 
@@ -157,7 +182,7 @@ final class Lexer
     {
         $position = $this->position;
 
-        while ($this->isNumber($this->char)) {
+        while ($this->isInteger($this->char) || $this->char === '.') {
             $this->advanceChar();
         }