+++ tests/LexerTest.php
@@ -77,6 +77,27 @@ class LexerTest extends TestCase
         ]);
     }
 
+    public function testLexingFloats(): void
+    {
+        $input = <<<HTML
+        <h1>{{ 2.5213 }} and {{ -1.3 }}</h1>
+        HTML;
+
+        $this->tokenizeString($input, [
+            new Token(TokenType::HTML, "<h1>"),
+            new Token(TokenType::OPENING_BRACES, "{{"),
+            new Token(TokenType::FLOAT, "2.5213"),
+            new Token(TokenType::CLOSING_BRACES, "}}"),
+            new Token(TokenType::HTML, " and "),
+            new Token(TokenType::OPENING_BRACES, "{{"),
+            new Token(TokenType::MINUS, "-"),
+            new Token(TokenType::FLOAT, "1.3"),
+            new Token(TokenType::CLOSING_BRACES, "}}"),
+            new Token(TokenType::HTML, "</h1>"),
+            new Token(TokenType::EOF, ""),
+        ]);
+    }
+
     public function testLexingBooleans(): void
     {
         $input = <<<HTML
@@ -293,6 +314,28 @@ class LexerTest extends TestCase
             new Token(TokenType::STRING, "Adult"),
             new Token(TokenType::COLON, ":"),
             new Token(TokenType::STRING, "Child"),
+            new Token(TokenType::CLOSING_BRACES, "}}"),
+            new Token(TokenType::EOF, ""),
+        ]);
+    }
+
+    public function testLexingIllegalTokens(): void
+    {
+        $input = <<<HTML
+        {{ 2.3.4 @ $ % ^ & * ( ) }}
+        HTML;
+
+        $this->tokenizeString($input, [
+            new Token(TokenType::OPENING_BRACES, "{{"),
+            new Token(TokenType::ILLEGAL, "2.3.4"),
+            new Token(TokenType::ILLEGAL, "@"),
+            new Token(TokenType::ILLEGAL, "$"),
+            new Token(TokenType::ILLEGAL, "%"),
+            new Token(TokenType::ILLEGAL, "^"),
+            new Token(TokenType::ILLEGAL, "&"),
+            new Token(TokenType::ILLEGAL, "*"),
+            new Token(TokenType::ILLEGAL, "("),
+            new Token(TokenType::ILLEGAL, ")"),
             new Token(TokenType::CLOSING_BRACES, "}}"),
             new Token(TokenType::EOF, ""),
         ]);